+++ tests/core/tests/test_admin_integration.py
@@ -680,7 +680,10 @@ class ExportAdminIntegrationTest(AdminTestMixin, TestCase):
             "file_format": "0",
         }
         date_str = datetime.now().strftime("%Y-%m-%d")
-        response = self.client.post("/admin/core/book/export/", data)
+        with self.assertNumQueries(
+            7
+        ):  # Should not contain COUNT queries from ModelAdmin.get_results()
+            response = self.client.post("/admin/core/book/export/", data)
         self.assertEqual(response.status_code, 200)
         self.assertTrue(response.has_header("Content-Disposition"))
         self.assertEqual(response["Content-Type"], "text/csv")